Generate Basic auth data
When a server offers Basic authentication, check if we have a callback
that we can call to retrieve the user name and password that we should
supply.

Cache those answers so that we don't prompt more than once, even if a
given command-line invocation results in several connections and
requests to the daemon.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com> (github: nalind)