Add an authentication framework
Add the basic framework for allowing clients to authenticate to the
server when issuing requests to it.  This patch is mainly plumbing and
wiring in the logic.

The client contains a module-local list of functions which create and
return handlers for authentication schemes.  The implementations add
themselves to this list at initialization time.  The AuthResponder
objects which they create are indexed in a map by the authentication
schemes that they implement.

The client will call the AuthResponder for a particular scheme if it
receives a 401 error, and if the AuthResponder indicates that the client
should try again, it will do so, repeating the process until it receives
something other than a 401 error from the server.  At that point, if the
request succeeded, it will call the scheme's AuthCompleted callback to
check if the server's success reply is acceptable to it.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com> (github: nalind)

Fixed a bug which disabled persistent connections when docker authentication is required

Signed-off-by: lior <lior@twistlock.com>