Switch from Rewinder to a full copy
Switch from attempting to use a Rewinder to buffering a full copy of the
client document at the start and creating new readers for it when we
retry.  Trying to use a rewinder wasn't saving us from ending up with a
full copy of the client data in memory in the success cases.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com> (github: nalind)