Correctly set content type for image load
The default content type is `text/plain`, which, if authz plugins are in
use will cause the entire body to be read (and fail because of another
bug that will be fixed in docker/docker separately). Given the content
type should really be binary format anyway, this is a good change for
any other code in the future that might determine action based on POST
content-type headers.

Docker-DCO-1.1-Signed-off-by: Phil Estes <estesp@linux.vnet.ibm.com> (github: estesp)