Implement Jenkins jobs using DSL
This creates a set of jobs to test this repo in various ways: vet, lint,
fmt, cyclo, unittests (using gocov test) on linux and windows. It also
includes jobs to trigger certain git refs: tags, branches, PRs, and the
default branch (master).

Cyclomatic complexity max is set to 21 and should be reduced over time as
functions are refactored.

vet, lint, fmt, and cyclo are all recorded using the Warnings plugin and
are currently set to fail the job if any issues are detected.

Test and coverage results are recorded in a jenkins-friendly fashion
using junit and cobertura plugins respectively.

Signed-off-by: Mike Dougherty <mike.dougherty@docker.com>