Initial commit for Jenkinsfile
This is a Jenkinsfile, used by the new(ish) Pipelines suite of plugins
that will be part of the defaults in Jenkins 2.0. I've reimplemented
most of the functionality in #136 using this instead. While it appears
to be less customizable, much of the "basic" functionality that all jobs
will need is more simply hidden away. Integration with GitHub is also
more complete and automatic.

Missing functionality: code coverage publishing

Signed-off-by: Mike Dougherty <mike.dougherty@docker.com>