Remove unnecessary strslice functions
strslice is intended to behave like a string slice with the additional property that when unmarshalled it can handle a string or array.
All over functions are unneeded and allow for misuse of the type.
Currently when output a strslice `fmt.Print` ends up getting used calling the `String()` method which has output which differs from the expected slice output.

Signed-off-by: Derek McGowan <derek@mcgstyle.net> (github: dmcgowan)