client: remove need for "test" tag to run tests
Because client tests imported a test file from another package, `go
test` would not work without adding `-tags=test`. This is not the
correct invocation for regularly testing Go code and breaks simple runs
of `go test` working for the package. This moves the transport mock into
the package that actually uses it. Now, `go test` works properly in the
client package.

Signed-off-by: Stephen J Day <stephen.day@docker.com>