types: stats: add pids.max to PidsStats
In order to allow nice usage statistics (in terms of percentages and
other such data), add the value of pids.max to the PidsStats struct
returned from the pids cgroup controller.

It has to be an int64 in order for "max" to be represented as -1 (which
mirrors the OCI specification for Resources.Pids.Limit).

Signed-off-by: Aleksa Sarai <asarai@suse.de>