Swagger spec for Docker APIs
Added a swagger spec for the Docker APIs.  This spec originated
from php-docker, massage by go-swagger, and fixed by vmware/vic.
Also added targets in the makefile to generate the swagger
client/server, build the client, and also to clean all the
generated code that gets recreated on regeneration cycles.  The
build is based on go-swagger.  The spec was tested against
go-swagger and the main swagger generator.

Signed-off-by: Loc Nguyen <nloc@vmware.com>