Move Docker version comparisons to the types package.
Most of the comparisons are done at the API level.
In the long term, it makes more sense to have these functions
in this package, so people can reuse them implementing other clients.

This change removes the custom Version type. The custom type doesn't
provide any extra value since you can wrap any string in it an the
comparisons can easily be done with plain strings.

Signed-off-by: David Calavera <david.calavera@gmail.com>