Use filters.ToParamWithVersion elsewhere
ToParamWithVersion allows to serialize correctly filters for older
version of docker (1.9.1 and lower mainly). It is currently only used
for ContainerList but it's not the only place where the filters happens.

This make use of ToParamWithVersion for ImageList, VolumesList,
NetworkList and Events as well (all where filters where already there in
1.9.1).

Signed-off-by: Vincent Demeester <vincent@sbr.pm>