Add config parameter to change per-container stop timeout during daemon shutdown
This fix tries to add a `StopTimeout` in container's Config so that a per-container
stop timeout could be specified during daemon shutdown.

The `StopTimeout` is defined as `*int` so that unspecified value could be
correctly identified as using the default value (10s).
Note: Both `0` and `-1` could not be used as the default value because
`0` means stop immediately and `-1` means wait forever.

This fix is related to docker:
https://github.com/docker/docker/pull/22566

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>