Add config parameter to change per-container stop timeout during daemon shutdown
This fix tries to add config paramter to change per-contaienr stop timeout during
the daemon shutdown. As the part of the update, this fix also changes `docker restart`
and `docker stop` so that if `--time` is not specified, then the default stop timeout
(10s) will be used.

This fix is related to
https://github.com/docker/docker/pull/22566
https://github.com/docker/docker/issues/22471

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>