Add `--force` in `docker volume rm` to fix out-of-band volume driver deletion
This fix tries to address the issue raised in

https://github.com/docker/docker/issues/23367

where an out-of-band volume driver deletion leaves some data in docker.
This prevent the reuse of deleted volume names (by out-of-band volume driver
like flocker).

This fix adds a `--force` field in `docker volume rm` to forcefully purge
the data of the volume that has already been deleted.

This fix is related to pull request in docker:

https://github.com/docker/docker/pull/23436

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>