Correct order of error checking on connect failure
1730fe8fb0452def2eafed4142bbd9139bda35db changed error checking to use
the `Temporary()` function from `net.Error`. Unfortunately having this
check so high up it is catching more than we really want it to and
causing the custom error checks below it to not be caught.

This moves the net.Error check to last since it is the most generic
check.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>