Change IPAM driver type to pointer to avoid default value
This fix tries to address the issue raised in 368 and docker/25735
where NetworkCreate needs to set up the IPAM driver type. Otherwise
an error will be returned.

This fix change IPAM driver type to pointer so that `/networks/create`
does not the explicit specification of IPAM driver any more.

This fix fixes 368.

This fix is related to docker/25735.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>